Rename and reorganize some pieces
