+++ fauna/v10/queries/remove.js
