+++ fauna/v4/RemoveQuery.js
